+++ src/main/java/com/zea7ot/lc/tags/algorithm/n_pointers/linear_scan.txt
@@ -2,6 +2,7 @@
 References:
     https://github.com/huanminwu/LeetCode/blob/master/LeetCode.xlsx
 
+
 ### N Pointers
 # Linear Scan / 1 Pointer
