+++ src/main/java/com/zea7ot/lc/tags/algorithm/n_pointers/summary.txt
@@ -6,6 +6,7 @@ References:
 ### Two Pointers
 # Summary
 
+1   0067    https://leetcode.com/problems/add-binary/
 1   0125    https://leetcode.com/problems/valid-palindrome/
 1   0345    https://leetcode.com/problems/reverse-vowels-of-a-string/
 1   0643    https://leetcode.com/problems/maximum-average-subarray-i/
