+++ src/main/java/com/zea7ot/lc/tags/data_structure/string/n_pointers/2_pointers/summary.txt
@@ -4,8 +4,10 @@ References:
 
 
 ### String
+## 2 Pointers
 # Summary
 
+1   0067    https://leetcode.com/problems/add-binary/
 1   0125    https://leetcode.com/problems/valid-palindrome/
 1   0345    https://leetcode.com/problems/reverse-vowels-of-a-string/
